modal

+++ user/gerenciamento_arvores.php
@@ -93,7 +93,7 @@ include('seguranca.php');
                             <td><?php echo $informacaoArvore->Situacao; ?></td>
                             <td>
                                 <a href="formulario.php?id=<?php echo $informacaoArvore->IdArvore; ?>"> Editar </a>
-                                <a href="excluirArvore.php?id=<?php echo $informacaoArvore->IdArvore; ?>"> Excluir </a>
+                                <a href="excluirArvore.php?id=<?php echo $informacaoArvore->IdArvore; ?>" data-confirm='Tem certeza de que deseja excluir o item selecionado?'> Excluir </a>
                                 <a href="formulario.php?id=<?php echo $informacaoArvore->IdArvore; ?>"> Ver </a>
                                 <a href="solicitar_servico.php?id=<?php echo $informacaoArvore->IdArvore; ?>"> Solicitar Serviço </a>
                             </td>
@@ -155,6 +155,7 @@ include('seguranca.php');
     <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
+    <script src="../js/modal_excluir.js"></script>
 </body>
 
 </html>
